backend/requirements.txt
Flask
Flask-PyMongo
python-dotenv
google-generativeai
chromadb
langchain
langchain-google-genai